1. Fixed a syntax error in oxleycpp.cpp 2. Changed the name of a function in RefinementZone to make it consistant with the rest of the codebase

+++ oxley/src/RefinementZone.cpp
@@ -37,7 +37,7 @@ RefinementZone2D::~RefinementZone2D()
 
 };
 
-void RefinementZone2D::refinePoint(double x0, double y0)
+void RefinementZone2D::refinePoint(float x0, float y0)
 {
 	Point2DRefinement refine(x0,y0);
 	Point2DRefinement * pRefine = &refine;
@@ -46,7 +46,7 @@ void RefinementZone2D::refinePoint(double x0, double y0)
 	addToQueue(*alg);
 };
 
-void RefinementZone2D::refineRegion(double x0, double y0, double x1, double y1)
+void RefinementZone2D::refineRegion(float x0, float y0, float x1, float y1)
 {
 	Region2DRefinement refine(x0,y0,x1,y1);
 	Region2DRefinement * pRefine = &refine;
@@ -55,7 +55,7 @@ void RefinementZone2D::refineRegion(double x0, double y0, double x1, double y1)
 	addToQueue(*alg);
 };
 
-void RefinementZone2D::refineCircle(double x0, double y0, double r)
+void RefinementZone2D::refineCircle(float x0, float y0, float r)
 {
 	CircleRefinement refine(x0,y0,r);
 	CircleRefinement * pRefine = &refine;
@@ -64,7 +64,7 @@ void RefinementZone2D::refineCircle(double x0, double y0, double r)
 	addToQueue(*alg);
 };
 
-void RefinementZone2D::refineBorder(Border b, double dx)
+void RefinementZone2D::refineBorder(Border b, float dx)
 {
 	Border2DRefinement refine(b,dx);
 	Border2DRefinement * pRefine = &refine;
@@ -83,7 +83,7 @@ RefinementZone3D::~RefinementZone3D()
 
 };
 
-void RefinementZone3D::refinePoint(double x0, double y0, double z0)
+void RefinementZone3D::refinePoint(float x0, float y0, float z0)
 {
 	Point3DRefinement refine(x0,y0,z0);
 	Point3DRefinement * pRefine = &refine;
@@ -92,7 +92,7 @@ void RefinementZone3D::refinePoint(double x0, double y0, double z0)
 	addToQueue(*alg);
 };
 
-void RefinementZone3D::refineRegion(double x0, double y0, double z0, double x1, double y1, double z1)
+void RefinementZone3D::refineRegion(float x0, float y0, float z0, float x1, float y1, float z1)
 {
 	Region3DRefinement refine(x0,y0,z0,x1,y1,z1);
 	Region3DRefinement * pRefine = &refine;
@@ -101,7 +101,7 @@ void RefinementZone3D::refineRegion(double x0, double y0, double z0, double x1, 
 	addToQueue(*alg);
 };
 
-void RefinementZone3D::refineSphere(double x0, double y0, double z0, double r)
+void RefinementZone3D::refineSphere(float x0, float y0, float z0, float r)
 {
 	SphereRefinement refine(x0,y0,z0,r);
 	SphereRefinement * pRefine = &refine;
@@ -110,7 +110,7 @@ void RefinementZone3D::refineSphere(double x0, double y0, double z0, double r)
 	addToQueue(*alg);
 };
 
-void RefinementZone3D::refineBorder(Border b, double dx)
+void RefinementZone3D::refineBorder(Border b, float dx)
 {
 	Border3DRefinement refine(b,dx);
 	Border3DRefinement * pRefine = &refine;
